Fix imports and get game to run
