+++ main.py
@@ -9,3 +9,4 @@
 from src.spades_env import SpadesEnv
 
 env = SpadesEnv()
+env.play_game()
